Refactor indexing-related fixtures
