+++ src/encoded/tests/test_validation_errors.py
@@ -1,5 +1,5 @@
 import pytest
-from .features.conftest import app_settings, workbook
+from .workbook_fixtures import app_settings, workbook
 from .test_search import delay_rerun
 pytestmark = [pytest.mark.working, pytest.mark.indexing] #pytest.mark.flaky(rerun_filter=delay_rerun)]
